refactor(mespapiers): Move makeInputAdornment function to helpers file
